Fix issue with Cassandra configured with non-routable `rpc_address`.
Cassandra documentation states that `rpc_address` can be set to 0.0.0.0 (or any
other non-routable value from the client point of view), as long as
`broadcadt_rpc_address` contains an address that clients can use to contact the
node.

This commit changes the lookup to `system.local` from fetching rpc_address to
broadcast_address.